refactor(node)!: Also move `Npm` into its own dedicated directory
Establish consistency and prepare for an upcoming refactoring of `Npm`
which will add further classes under the `npm` directory.
