+++ plugins/package-managers/node/src/funTest/kotlin/npm/NpmFunTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.plugins.packagemanagers.node
+package org.ossreviewtoolkit.plugins.packagemanagers.node.npm
 
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.engine.spec.tempdir
