+++ plugins/package-managers/node/src/test/kotlin/npm/NpmTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.plugins.packagemanagers.node
+package org.ossreviewtoolkit.plugins.packagemanagers.node.npm
 
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.engine.spec.tempdir
